model/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.5.2
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
tqdm==4.66.2
